script to verify the value of PCR10
modification of the custom template "ima-id" to have the fields as strings. this was done because are easier to handle during the attestation.

Insertion of the script to check the value of the PCR10, simulates the extension of the PCR for each entry of the "ascii_runtime_measurements"
usage: execute as root and give as parameter the current value of the PCR10, stored in /sys/class/tpm/tpmX/prc-sha1/10,
the script reconstruct the value of the hash template of the measure done on child namespaces.