IMA namespace ID
Insertion of a unique identifier of the IMA namespace. The identifier is incremental for each new namespace created.
Host's IMA namespace will have ID 1, the namespace created after will have 2...

The idea behind the ID is to identify the IMA namespace that started a measure, useful when extending each event to
the partent up to the host.