IMA namespace creation/destroy event measure
The creation and destroy of a IMA namespace is recorded in the measurement list.
The template "ns-event" is used with the fields:
- imaeveninfo to specify if is a creation (the field will have value 0) or a destruction (the field will have value 1)
- imaidcreator that indicates the id of the IMA namespace of the creator
- imaidcreated that indicated the id of the IMA namespace of the created

All the creation/destruction events are recorded in every measurement list of all the parents in chain up to the host.

The function ima_ns_event in ima_api.c handles the allocation of all the structures needed, and calls the store_template
to save the value in the list. Since there is no policy regarding this measure all needs to be allocated here.
This extension is mandatory in case of IMA namespace enabled. This entry may be used to verify the list.

Problems: reconstruct the chain in case of longer than 2 chains (if the namespaces are not directly created by the host).