Insertion of the verification script + Bugfixes
The script that contains the logic to implement a tree to reconstruct the number of extensions done in the PCR for eache entry.
The library tree_lib contains the class for handle the tree creation
Bugfixes of the show of IMA namespace identifier in the entry, now shown correctly.
Bugfix on the enxtension of the measure done by the parent chain. In past was possible that a measure coming from
another namespace extends the PCR while another namespace is performing parent extension. Since the integrity
verification method is based on replicate the extension made by childs the extension related to a measure have
to be atomic. Now a lock is inserted and this behaviour is impossible to happen anymore.