Insertion of a queue for ima measures
To handle the problem that a measure extension is not atomic, in this case a vector is implemented to ordinate the
measure time of the data.
The vector is based on the ima ID and is better than just a lock because since is a FIFO the wait time have a max.
when a event starts his id is inserted in the vector, when the pointer in the vector that indicates the ID of the
namespace that is able to measure is equal to the current namespace it will be able to insert value in the list
and perform PCRs extensions