Implementation of the vPCR logic
Insertion of a function "host_extension_vpcr".
Every time a namespace (not the host one) perform
a measure it will extend his virtual PCR and
send to the host namespace the current value of
his vPCR. The host will save in his list the value
of the vPCR and extend the true PCR10.

To verify this solution is possible to retrive value of the
measurement list of the namespace under verification.
Firstly is needed to verify the integrity of the host list
then is possible to extract the last value of the vPCR of the
namespace starting from the list of the host.
Reconstruct the vPCR starting from the namespaces' list and
check if correspond with the value stored in the host's list.

The new template entry will have the fields regarding
- vPCR new value
- identifier of the ima namespace that started the measure